# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskForge API",
    "version": "1.0.0",
    "description": "Authentication-enabled API for TaskForge. Use the `/auth/login` or `/auth/register` endpoints to obtain a JWT before calling protected routes."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tf_session",
        "description": "HttpOnly session cookie issued by the auth routes. The cookie carries the same JWT used for bearer authentication."
      },
      "sessionBridgeSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-session-bridge-secret",
        "description": "Shared server-side secret used by the web app to exchange a NextAuth session for API JWTs."
      },
      "jobSecretHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-job-secret",
        "description": "Shared secret required by protected background job routes."
      },
      "jobSecretBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shared-secret",
        "description": "Alternative protected job authentication using `Authorization: Bearer <DIGEST_JOB_SECRET>`."
      }
    },
    "schemas": {
...
      "AuthMeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "email",
              "createdAt"
            ],
            "example": {
              "id": "26f26639-05ad-40b6-8248-379644dcdf18",
              "email": "user@example.com",
              "createdAt": "2024-06-01T12:00:00.000Z"
            }
          }
        },
        "required": [
          "user"
        ],
        "example": {
          "user": {
            "id": "26f26639-05ad-40b6-8248-379644dcdf18",
            "email": "user@example.com",
            "createdAt": "2024-06-01T12:00:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "description": "Optional structured validation details",
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Invalid payload",
          "details": {
            "fieldErrors": {
              "email": [
                "Invalid email address"
              ]
            },
            "formErrors": []
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/taskforge/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve the authenticated user",
        "description": "Requires a valid JWT provided via the `Authorization: Bearer <token>` header or the `tf_session` HttpOnly cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                },
                "examples": {
                  "authenticated": {
                    "value": {
                      "user": {
                        "id": "26f26639-05ad-40b6-8248-379644dcdf18",
                        "email": "user@example.com",
                        "createdAt": "2024-06-01T12:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/taskforge/v1/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve the authenticated user and account preferences",
        "description": "Requires a valid JWT provided via the `Authorization: Bearer <token>` header or the `tf_session` HttpOnly cookie. Includes daily digest email preferences for signed-in users.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user and email preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMeResponse"
                },
                "examples": {
                  "authenticated": {
                    "value": {
                      "user": {
                        "id": "26f26639-05ad-40b6-8248-379644dcdf18",
                        "email": "user@example.com",
                        "createdAt": "2024-06-01T12:00:00.000Z"
                      },
                      "emailPreference": {
                        "dailyDigestEnabled": false,
                        "dailyDigestTimezone": "UTC"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/taskforge/v1/me/email-preferences": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Update email preferences for the authenticated user",
        "description": "Enables or disables daily digest emails for the signed-in user. Requires a valid JWT provided via the `Authorization: Bearer <token>` header or the `tf_session` HttpOnly cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailPreferenceUpdateInput"
              },
              "examples": {
                "default": {
                  "value": {
                    "dailyDigestEnabled": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email preferences updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPreferenceResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "emailPreference": {
                        "dailyDigestEnabled": false,
                        "dailyDigestTimezone": "UTC"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalid": {
                    "value": {
                      "error": "Invalid payload",
                      "details": {
                        "fieldErrors": {
                          "email": [
                            "Invalid email address"
                          ]
                        },
                        "formErrors": []
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/taskforge/v1/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List tasks for the authenticated user",
        "description": "Returns a paginated collection of the signed-in user's tasks. Requires a valid JWT via `Authorization` header or the `tf_session` cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number (1-indexed)."
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of tasks per page."
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "TODO",
                "IN_PROGRESS",
                "DONE"
              ]
            },
            "description": "Filter tasks by workflow status."
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
              ]
            },
            "description": "Filter tasks by priority."
          },
          {
            "name": "tag",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "maxItems": 10,
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 64
              }
            },
            "description": "Filter tasks that include the specified tag(s). Repeat the parameter to require multiple tags."
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "description": "Case-insensitive search over the title and description."
          },
          {
            "name": "dueFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Only return tasks due on or after this ISO timestamp."
          },
          {
            "name": "dueTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Only return tasks due on or before this ISO timestamp."
          }
        ],
        "responses": {
          "200": {
            "description": "Task collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "items": [
                        {
                          "id": "9e22c508-1383-4609-9bbd-2e09b7a2d108",
                          "title": "Draft project brief",
                          "description": "Summarize goals and milestones for the release",
                          "status": "IN_PROGRESS",
                          "priority": "HIGH",
                          "dueDate": "2024-07-10T16:00:00.000Z",
                          "tags": [
                            "planning",
                            "product"
                          ],
                          "createdAt": "2024-06-01T12:00:00.000Z",
                          "updatedAt": "2024-06-03T09:30:00.000Z"
                        },
                        {
                          "id": "14377d29-0a0f-4b26-8b6d-60406ad3f7c1",
                          "title": "Follow up with design partners",
                          "description": "Confirm handoff expectations and surface risks early.",
                          "status": "TODO",
                          "priority": "MEDIUM",
                          "dueDate": "2024-07-18T15:00:00.000Z",
                          "tags": [
                            "customer",
                            "outreach"
                          ],
                          "createdAt": "2024-06-02T14:22:00.000Z",
                          "updatedAt": "2024-06-04T11:10:00.000Z"
                        }
                      ],
                      "page": 1,
                      "pageSize": 20,
                      "total": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidFilters": {
                    "value": {
                      "error": "Invalid payload",
                      "details": {
                        "fieldErrors": {
                          "dueFrom": [
                            "dueFrom must be earlier than or equal to dueTo"
                          ]
                        },
                        "formErrors": []
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a task",
        "description": "Creates a new task owned by the authenticated user. Requires a valid JWT via `Authorization` header or the `tf_session` cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateInput"
              },
              "examples": {
                "default": {
                  "value": {
                    "title": "Book product sync",
                    "description": "Coordinate roadmap review with stakeholders",
                    "status": "TODO",
                    "priority": "HIGH",
                    "tags": [
                      "planning"
                    ],
                    "dueDate": "2024-07-05T15:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                },
                "examples": {
                  "default": {
                    "value": {
                      "id": "9e22c508-1383-4609-9bbd-2e09b7a2d108",
                      "title": "Draft project brief",
                      "description": "Summarize goals and milestones for the release",
                      "status": "IN_PROGRESS",
                      "priority": "HIGH",
                      "dueDate": "2024-07-10T16:00:00.000Z",
                      "tags": [
                        "planning",
                        "product"
                      ],
                      "createdAt": "2024-06-01T12:00:00.000Z",
                      "updatedAt": "2024-06-03T09:30:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalid": {
                    "value": {
                      "error": "Invalid payload",
                      "details": {
                        "fieldErrors": {
                          "email": [
                            "Invalid email address"
                          ]
                        },
                        "formErrors": []
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/taskforge/v1/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "List tags",
        "description": "Retrieves all tag labels created by the authenticated user. Requires a valid JWT via `Authorization` header or the `tf_session` cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tag collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagListResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "items": [
                        {
                          "label": "planning",
                          "count": 3
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Create a tag",
        "description": "Creates a new tag for the authenticated user. Requires a valid JWT via `Authorization` header or the `tf_session` cookie.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateInput"
              },
              "examples": {
                "default": {
                  "value": {
                    "label": "Planning"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRecord"
                },
                "examples": {
                  "default": {
                    "value": {
                      "id": "8d367c45-8d3c-4e96-a59e-4254920fb828",
                      "label": "planning"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "value": {
                      "error": "Too many requests, please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "security": []
}